+++ model/src/main/kotlin/utils/RootLicenseMatcher.kt
@@ -24,7 +24,7 @@ import java.io.File
 import org.ossreviewtoolkit.model.LicenseFinding
 import org.ossreviewtoolkit.model.config.LicenseFilenamePatterns
 import org.ossreviewtoolkit.utils.common.FileMatcher
-import org.ossreviewtoolkit.utils.core.getAllAncestorDirectories
+import org.ossreviewtoolkit.utils.common.getAllAncestorDirectories
 
 /**
  * A heuristic for determining which (root) license files apply to any file or directory.
